chore: add initial code to convert form themes to class
