--- src/Views/Form-Themes/Sequoia/functions.php
@@ -1,4 +1,0 @@
-<?php
-require_once 'actions.php';
-require_once 'filters.php';
-require_once 'scripts.php';
