+++ src/Views/Form/Themes/Sequoia/sections/progress-bar.php
